requirements.txt
pillow
openai
python-dotenv
retry